requirements.txt
Pillow
requests
numpy
matplotlib
mysql-connector